+++ src/main/java/calculator/view/OutputView.java
@@ -2,7 +2,7 @@ package calculator.view;
 
 public class OutputView {
 
-    public static void printResult(int result) {
+    public static void printResult(long result) {
         System.out.println("결과 : " + result);
     }
 }
